funcionamento da animação

+++ include/ui/Button.h
@@ -19,7 +19,7 @@ public:
     Button(float x, float y, float width, float height, const std::function<void()> &onClick);
 
     void handleEvent(const SDL_Event &event) override;
-    void update(float delta) override;
+    void update(double delta) override;
     void render(SDL_Renderer *renderer, SDL_Texture *texture) override;
 
     static SDL_Texture* textureGenerator(SDL_Renderer *renderer, const std::string &path);
